requirements.txt
flask==3.0.3
twilio==9.2.3
google-generativeai==0.7.2
reportlab==4.2.2
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
langdetect==1.0.9
gunicorn==22.0.0
groq==0.11.0
httpx==0.27.2
